+++ CHANGELOG.md
@@ -43,6 +43,8 @@
 - Enh #214: Improved interface hierarchy (@samdark)
 - Enh #207: More specific Psalm type for `OffsetPaginator::withCurrentPage()` parameter (@samdark)
 - Enh #210: More specific Psalm type for `PaginatorInterface::getPageSize()` result (@vjik)
+- Enh #219: Add ability to set custom message in `PageNotFoundException` (@vjik)
+- Enh #219: Throw `PageNotFoundException` instead of `PaginatorException` in `OfferPaginator::withCurrentPage()` (@vjik)
 
 ## 1.0.1 January 25, 2023